Still trying to get proposeNodes working.  Realised it's because I need to define equals(); but ...
before I do that, I realised that I need to be able to have (proposed) Edges which aren't in any Graph,
because the complete set of proposed edges might contain cycles, even if adding any one of them
to the Graph wouldn't cause a cycle.  This will affect how (or, where) I define equals(),
so let's do that first.