#5 Change order of proposed so the graph is likely to become more linear.
Plus, now you can click on either edge in a proposed pair to add an edge from that one to the other.

Also, I added a list of edges with buttons to remove them, but that doesn't work because of my weird
attempt to have polymorphic equality: the equals method might work but the hashCode doesn't, because
it includes the hash code of DocWrapper, which includes the _rev.  Time to re-think how I want to
connect DB docs to "in-memory" objects (i.e., make it not be a quick hack).