#5 Improve edge proposal; fix ranking; ... but poor performance.
Get edge removal to work by excluding the _rev from DocWrapper equals & hashCode.

If there's only one node at a given rank, don't propose adding edges from it to other nodes,
because we've already narrowed down that rank as much as we can.

If we're proposing A->B, don't also propose a lower-ranked B->A, because the UI allows users
to add the edge in either direction.

Fix rank calculation.  Previously it gave each node a rank equal to the first depth at which
it was encountered in a traversal, where the traversal stopped after visiting each node once.
The right answer is to give it the max depth when following all edges, even though that means
visiting nodes several times.

The app now has the basic behaviour I want (for a single axis) but the performance is
terrible once I have about 9 ranks.  I'm hoping that's because my rank calculation is re-done
from scratch every time an edge is added or removed, whereas you could recalculate it
incrementally.