#17 Generalise the idea of scoring some nodes.
Factor out the idea of scoring some nodes from the concrete implementation of
scoring by rank in a graph.  This will allow for scoring along axes such as
"urgency", which is something like "1 / days until deadline".

This currently doesn't update the overall list of scores because, when each
scoring (graph) is updated (has an edge added), the owner scoring needs to
be notified so that it can (lazily) update.  I'll fix that another day.