#20 Fix a misuse of inlining which harmed performance.
We were inlining a "function returning Unit" into a call to setTImeout.
Kotlin's type system allows this (given the definition of setTimeout) but it's
not what we want: the inlined function is executed before the call to
setTimeout, which is called with the Unit value (which is "undefined" in
JavaScript), so the rendering update wasn't lazy.

Also, removed a warning in jsobject().