#23 Prototype work for fetching tasks from Outlook.
This so far only manages to log in to Office 365 ... but it turns out the
on-premise Exchange server I care about is Exchange 2010, and would need to be
Exchange 2016 CU3 to be compatible.  I'll have to look into the SOAP API
instead, but I'll leave this here on a branch.